+++ shelf_packages/open_api_client_generator/CHANGELOG.md
@@ -1,5 +1,8 @@
+
 ## 1.1.0
 - chore: removed support to dart http client generator
+- build: bumped `analyzer` dependency to `>=7.4.0 <9.0.0`
+- build: allow `code_builder: ^4.4.0`
 
 ## 1.0.1
 - fix: fixed security field in open api operation